Address Copilot review feedback
- Move initLog inside try/catch so failures route through onError
- Use updateSessionStatus instead of updateSession for initial
  "running" transition to preserve runtimeSessionId
- Extract broadcast into ws-broadcast.ts to break circular dependency
  between event-processor.ts and ws-bridge.ts
